move input generators
