+++ optimum/utils/import_utils.py
@@ -22,6 +22,7 @@ from typing import Tuple, Union
 
 import numpy as np
 from packaging import version
+
 from transformers.utils import is_torch_available
